+++ swift_browser_ui_frontend/src/components/ObjectTable.vue
@@ -2,19 +2,7 @@
   <div
     id="object-table"
   >
-    <c-row>
-      <router-link
-        class="back-link"
-        :to="getRedirectRoutes()"
-      >
-        <i class="mdi mdi-chevron-left" />
-        {{ isSharingFolder ? $t("message.table.back_to_sharing_folders")
-          : isSharedFolder ? $t("message.table.back_to_shared_folders")
-            : $t("message.table.back_to_all_folders")
-        }}
-      </router-link>
-    </c-row>
-
+    <BreadcrumbNav />
     <div class="folder-info">
       <div class="folder-info-heading">
         <i class="mdi mdi-folder-outline" />
@@ -150,13 +138,14 @@ import { useObservable } from "@vueuse/rxjs";
 import CObjectTable from "@/components/CObjectTable.vue";
 import debounce from "lodash/debounce";
 import escapeRegExp from "lodash/escapeRegExp";
-
+import BreadcrumbNav from "@/components/BreadcrumbNav.vue";
 import { toRaw } from "vue";
 
 export default {
   name: "ObjectTable",
   components: {
     CObjectTable,
+    BreadcrumbNav,
   },
   filters: {
     truncate,
@@ -220,6 +209,11 @@ export default {
       await this.getFolderSharedStatus();
       await this.updateObjects();
     },
+    client: async function() {
+      await this.getSharedContainers();
+      await this.getFolderSharedStatus();
+      await this.updateObjects();
+    },
     searchQuery: function () {
       // Run debounced search every time the search box input changes
       this.debounceFilter();
@@ -279,26 +273,6 @@ export default {
     this.abortController.abort();
   },
   methods: {
-    getRedirectRoutes: function () {
-      if (this.isSharingFolder) {
-        return {
-          name: "SharedFrom",
-          params: { project: this.$store.state.active.id },
-        };
-      }
-      else if (this.isSharedFolder) {
-        return {
-          name: "SharedTo",
-          params: { project: this.$store.state.active.id },
-        };
-      }
-      else {
-        return {
-          name: "AllFolders",
-          params: { project: this.$store.state.active.id },
-        };
-      }
-    },
     getSharedContainers: async function () {
       this.sharedContainers = await getSharedContainers(this.active.id);
     },
@@ -630,18 +604,6 @@ export default {
 
 #search {
   flex: 0.4;
-}
-
-.back-link {
-  display: flex;
-  padding-bottom: .5rem;
-  color: $csc-primary;
-  font-weight: 600;
-  align-items: center;
-
-  & .mdi {
-    font-size: 2rem;
-  }
 }
 
 .folder-info {